+++ LuckyoneUI/Init.lua
@@ -67,6 +67,7 @@ Private.isRetail = WOW_PROJECT_ID == WOW_PROJECT_MAINLINE
 
 -- API checks
 Private.IsAddOnLoaded = IsAddOnLoaded
+Private.IsEditModeActive = Private.isRetail or Private.isMists or Private.isTBC
 Private.Version = tonumber(GetAddOnMetadata(Name, 'Version'))
 
 -- Player utils