+++ CHANGELOG.md
@@ -1,4 +1,24 @@
-﻿## Version 4.14 [22-05-2026]
+﻿## Version 4.15 [03-06-2026]
+
+**All WoW Versions:**  
+- ElvUI 15.15+ compatibility  
+- Disabled Action Bar 15 by default (Its still configured and can be re-enabled)  
+- Reworked Debuffs on Party and Raidframes  
+- Updated ruRU localization (Credits: Hollicsh)  
+
+**Retail:**  
+- Added external defensives to player unitframe by default (Centered)  
+- Added option to select a different ElvUI target nameplate texture  
+- Reworked Private Auras on Party and Raidframes  
+- Updated Northern Sky Raid Tools profile  
+- Updated SkironCooldownManager profile (Requires alpha)  
+
+**Mists:**  
+- Added addonlist icon  
+- Edit mode support  
+- Updated TOC for 5.5.4  
+
+## Version 4.14 [22-05-2026]
 
 **All WoW Versions:**  
 - ElvUI 15.14+ compatibility